requirements.txt
asgiref==3.5.2
Django==4.0.6
django-birthday==0.1.4
docopt==0.6.2
num2words==0.5.12
numpy==1.23.2
Pillow==9.2.0
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
